--- packages/app-project/src/screens/ProjectSubjectPage/index.js
@@ -1,1 +1,0 @@
-export { default } from './ProjectSubjectPageConnector'
