+++ packages/app-project/src/screens/SubjectTalkPage/index.js
@@ -1,0 +1,1 @@
+export { default } from './SubjectTalkPageConnector'
